rename folder and change auth method
