+++ src/skills/shortcut.js
@@ -1,7 +1,7 @@
 import { say, chatEvents } from '../components/chatClient.js';
 import { logger } from '../logger/index.js';
-import { load } from '../service/configFiles.js';
-import { isUserStreaming } from '../service/stream.js';
+import { load } from '../services/configFiles.js';
+import { isUserStreaming } from '../services/stream.js';
 
 export const shortcuts = () => {
   try {
